Move Preferences to interfaces sub-module.
